+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0
+
+* Add the `PiPStatusInfo` class and add the `isCreateNewEngine` and `isEnabledWhenBackground` for
+  the current pip
+* Fixed `disable()` not working in android when `createNewEngine=true`
+* Change the `isActive()` return parameter to `PiPStatusInfo`
+
 ## 1.0.0
 
 * Removed `FlPiP().enableWithEngine`